requirements.txt
asgiref==3.3.4
autopep8==1.5.6
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
cloudinary==1.28.0
dj-database-url==0.5.0
Django==3.2
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.1
django-extensions==3.1.3
django-multiupload2==0.5.3
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
Pillow==8.2.0
psycopg2-binary==2.9.3
pycodestyle==2.7.0
pytz==2021.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.8
whitenoise==5.3.0
